// src/models/Store.ts
export interface Store {
  id: string;
  image: string;
  title: string;
  businessHour?: string;
  detail: string;
  place: string;
  phonenumber?: string;
  email: string;
  homepage?: string;
  twitter?: string;
  instagram?: string;
  facebook?: string;
  tabelog?: string;
  otherUrl?: string;
  latitude: number;
  longitude: number;
}